NEM-90 / design : FoodItem 컴포넌트 추가
- FoodItem 컴포넌트 추가
- FoodCatrgoryItems state에 따라서 스타일 변경 완성
